@@ -8,22 +8,21 @@
     "get-js": "curl \"http://localhost:8081/index.android.bundle?platform=android&dev=false&minify=true\" -o \"android/app/src/main/assets/index.android.bundle\"",
 	"genkey": "keytool -genkey -v -keystore scs.keystore -alias scs -keyalg RSA -keysize 2048 -validity 10000",
     "release": "cd android && gradlew assembleRelease && cd ..",
-    "install-dbg": "adb install -r android/app/build/outputs/apk/appstarter-debug-1.0.apk",
-    "install-rel": "adb install -r android/app/build/outputs/apk/appstarter-release-1.0.apk",
+    "install-dbg": "adb -d install -r android/app/build/outputs/apk/appstarter-debug-1.0.apk",
+    "install-rel": "adb -d install -r android/app/build/outputs/apk/appstarter-release-1.0.apk",
     "bridge": "adb reverse tcp:8081 tcp:8081",
     "menu": "adb shell input keyevent 82",
     "logs": "adb logcat *:S ReactNative:V ReactNativeJS:V"
   },
   "dependencies": {
+    "app-nub.react": "git+ssh://git@github.com:jcapuano328/app-nub.react.git",
+    "dice.react": "git+ssh://git@github.com:jcapuano328/dice.react.git",
     "apsl-react-native-button": "^3.0.0",
-    "autobind-decorator": "^1.3.3",
     "moment": "^2.14.1",
-    "react": "15.2.1",
-    "react-native": "^0.31.0",
-    "react-native-audioplayer": "^0.2.0",
+    "react": "^15.3.2",
+    "react-native": "^0.34.0",
     "react-native-fs": "^1.5.1",
     "react-native-menu": "^0.20.1",
-    "react-native-scrollable-tab-view": "^0.5.4",
-    "react-native-simple-radio-button": "^2.2.0"
+    "react-native-scrollable-tab-view": "^0.5.4"
   }
 }
